chat_service/requirements.txt
Django>=4.2.0,<5.3.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-core>=0.2.0
langchain-neo4j>=0.1.0
psycopg2-binary>=2.9.9
